Kernel/SysFS: Split the bulky BIOS.h file into multiple files
